requirements.txt
# Bluetooth Mesh Broadcast Application - Dependencies
# Python 3.10+

# Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.1

# Async Support
eventlet==0.34.2
greenlet==3.0.1

# Bluetooth (BLE)
bleak==0.21.1

# Caching
cachetools==5.3.2

# Rate Limiting
Flask-Limiter==3.5.0
limits==3.7.0

# Environment Variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Resource Monitoring
psutil==5.9.7

# Validation
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
